+++ packages/@expo/cli/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### 🐛 Bug fixes
 
+- Don't print source map size in `npx expo export` when the source maps are not written.
+
 ### 💡 Others
 
 ## 0.4.0 — 2022-10-25